+++ src/Client.php
@@ -36,7 +36,7 @@ final class Client {
 	 * @var string The current drop-in file version
 	 * @since 1.0.0
 	 */
-	const VERSION = '1.0.1';
+	const VERSION = '1.0.2';
 
 	/**
 	 * @var Config
@@ -444,6 +444,7 @@ final class Client {
 		do_action( 'trustedlogin/' . $this->config->ns() . '/access/extended', array(
 			'url'    => get_site_url(),
 			'action' => 'extended',
+			'ref' => self::get_reference_id(),
 		) );
 
 		return $return_data;